requirements.txt
torch>=1.10.0
torchvision>=0.11.1
fastapi>=0.70.0
uvicorn>=0.15.0
python-multipart>=0.0.5
pillow>=8.4.0
boto3>=1.20.0
pytest>=6.2.5
requests>=2.26.0
opencv-python>=4.5.3
mangum>=0.12.0